Refactor Balance Fetching Modules
